Leonardo Heap implementeaton , TODO: Test cases
